+++ test_autofit/unit/config/json_priors/test_autofit.json
@@ -135,6 +135,37 @@
                     "upper": 1.0
                 }
             }
+        },
+        "Tracer": {
+            "lens_galaxy": {
+                "type": "Uniform",
+                "lower_limit": 0.0,
+                "upper_limit": 1.0,
+                "width_modifier": {
+                    "type": "Absolute",
+                    "value": 0.2
+                },
+                "gaussian_limits": {
+                    "lower": 0.0,
+                    "upper": 1.0
+                }
+            },
+            "source_galaxy": {
+                "type": "Uniform",
+                "lower_limit": 0.0,
+                "upper_limit": 1.0,
+                "width_modifier": {
+                    "type": "Absolute",
+                    "value": 0.2
+                },
+                "gaussian_limits": {
+                    "lower": 0.0,
+                    "upper": 1.0
+                }
+            },
+            "grid": {
+                "type": "Deferred"
+            }
         }
     }
 }
